adds auto detection features
As per the request of @raydog153 I added in the ability to manually
enter the TV's MAC address if necessary

Additions.
Auto detection for the MAC address
Auto detection for the connection method
Auto detection for the port

all three if these things were added to the samsungctl.Config class.
The Readme has also been updated.

KNOWN ISSUES
encrypted connection is currently broken. I am aware of this I am
just waiting for the author of the fix to let me know how to proceed
with adding the fix to the code.