massive update to the discovery. I moved all of the power handling and also the parsing of the SSDP packets into the individual threads for each of the adapter addresses. this makes it much easier to handle the power state of the TV's without it affecting any of the other threads.
I put into place a smoothing of sorts for the power state. I did this
so that if the TV did not manage to get the UPNP SSDP response out
before the loop closed it would not be considered as powered off right
out of the gate. so it gets placed into a "powering off" state. when
this happens I set the socket tiimeout to 10 seconds to ensure that the
packet gets received (or not) the second tme around. after the second
loop the timeout on the socket gets set back to 5. and if the TV did
not respond that it is considered powered off.

There is an issue with using WiFi on the TV's and then not disassociating
properly from an Access Point. Because of this when a TV gets powered
back on it causes a kind of network outage for the TV. II have tried
using a static IP and this did not sem to help. I know the issue exists
with legacy TV's I am not sure about any of the others. so the solution
at this point is to bee on a wired connection. Which is something you
should be doing for anything HA related anyway.