fixes several bugs
fixes incompatible libCEC version (generates error and program can continue to run)
should fix the shell / command prompt use of the library
I have also done a lot of modification to the power on and off portions
of the library. The issue I am running into the the amount of time it
takes for a TV to power off. I still will need to make more changes to it.
I do not want to stall the program waiting for the TV to power back on.
The reason is if the TV was just powered off this process could take
over 2 minutes before the TV fully powers off and back on again.
I am trying to find the solution to be able to trim this time down.
This problem seems to be only in the latest 2018+ samsung TV's.
The TV's will not acknowledge a WOL packet until it has been fully
powered off (not just the screen going black). I am trying to leverage the
websocket to undo the shutdown process. I do not think the KEY_POWERON
key is there for no reason. There is evidence of it stopping the power
off process, I do not know what models this happens with.